[gk_pr_3] Интерфейсы для работы со слоем service объявлены там где они используются
